Swagger

+++ core-service/src/main/java/com/prumi/web/core/entities/Product.java
@@ -6,6 +6,7 @@ import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -26,7 +27,7 @@ public class Product {
     private ProductCategory productCategory;
 
     @Column(name = "price")
-    private Integer price;
+    private BigDecimal price;
 
     @CreationTimestamp
     @Column(name = "created_at")
@@ -37,7 +38,7 @@ public class Product {
     private LocalDateTime updatedAt;
 
 
-    public Product(Long id, String title, ProductCategory productCategory, Integer price) {
+    public Product(Long id, String title, ProductCategory productCategory, BigDecimal price) {
         this.id = id;
         this.title = title;
         this.productCategory = productCategory;
